+++ SignalTester/plot/plot.hpp
@@ -10,7 +10,7 @@ class Plot {
 public:
     Plot(std::string name);
 
-    void writeToFile(std::vector<double> sig, double idxstep = 1.0, double idxoffset = 0.0);
+    void writeToFile(std::vector<float> sig, double idxstep = 1.0, double idxoffset = 0.0);
     void close();
 
 private:
